refactor admin add event form to CalendarEventForm component
